refactor convert
